requirements.txt
torch==2.6.0
numpy>=1.22.4,<2
pandas==1.5.2
accelerate==0.26.0
einops==0.6.1
diffusers==0.32.2
transformers==4.49.0
scipy==1.11.0
tqdm==4.65.0
python-dotenv==1.0.1
openai==1.59.3
lpips==0.1.4
ftfy==6.2.3
seaborn==0.13.2
matplotlib==3.8.4
opencv-python==4.8.1.78
pytorch-lightning==2.1.0
huggingface-hub==0.29.1
datasets==2.6.2
dash==2.3.1
peft==0.5.0
omegaconf==2.3.0
langchain==0.3.13
langchain-community==0.3.13
langchain-core==0.3.28
langchain-openai==0.2.10
langchain-text-splitters==0.3.4
langsmith==0.2.7
# for IP-Adapter
git+https://github.com/tencent-ailab/IP-Adapter.git
git+https://github.com/openai/CLIP.git